#3077 improve intellisense for variables in sass and less
